# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste prático - PHP",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/pedidos": {
            "get": {
                "tags": [
                    "pedido"
                ],
                "summary": "Listar pedidos",
                "description": "Retornar todos os pedidos",
                "operationId": "listar_pedidos",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pedido/{id}": {
            "get": {
                "tags": [
                    "pedido"
                ],
                "summary": "Detalhar pedido",
                "description": "Retornar todos os pedidos",
                "operationId": "detalhar_pedido",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "pedido"
                ],
                "summary": "Excluir pedido",
                "operationId": "excluir_pedido",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/relizar-pedido": {
            "post": {
                "tags": [
                    "pedido"
                ],
                "summary": "Realizar novo pedido",
                "operationId": "realizar_pedido",
                "requestBody": {
                    "description": "Campos para realizar o pedido",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "produtos"
                                ],
                                "properties": {
                                    "produtos": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "produto_id": {
                                                    "type": "integer"
                                                },
                                                "quantidade": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    },
                    "422": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pedido-item/{id}": {
            "delete": {
                "tags": [
                    "pedido"
                ],
                "summary": "Excluir item do pedido",
                "operationId": "excluir_item_pedido",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth"
            }
        }
    ]
}